chore(workspace): add root Cargo workspace (research/cartography.md §5.1-2/§5.6)

+++ sdk/rust/src/filesystem/agentfs.rs
@@ -8637,7 +8637,7 @@ mod tests {
             &fs,
             stats.ino,
             TimeChange::Omit,
-            TimeChange::Set(explicit_secs, explicit_nsec as u32),
+            TimeChange::Set(explicit_secs, explicit_nsec),
         )
         .await?;
